+++ Source/Kysect.Shreks.sln
@@ -80,6 +80,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.WebUI.AdminPa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.WebApi.Sdk", "Presentation\Kysect.Shreks.WebApi.Sdk\Kysect.Shreks.WebApi.Sdk.csproj", "{3B2120BB-437C-4485-A039-EF4C0CD5D94D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Tests.Handlers", "Tests\Kysect.Shreks.Tests.Handlers\Kysect.Shreks.Tests.Handlers.csproj", "{93EFB7C0-9A60-4C33-8415-863E5F4FA9A4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -198,6 +200,10 @@ Global
 		{3B2120BB-437C-4485-A039-EF4C0CD5D94D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3B2120BB-437C-4485-A039-EF4C0CD5D94D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3B2120BB-437C-4485-A039-EF4C0CD5D94D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{93EFB7C0-9A60-4C33-8415-863E5F4FA9A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93EFB7C0-9A60-4C33-8415-863E5F4FA9A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{93EFB7C0-9A60-4C33-8415-863E5F4FA9A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{93EFB7C0-9A60-4C33-8415-863E5F4FA9A4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -232,6 +238,7 @@ Global
 		{356898AB-4F8A-43C4-945C-670B2757848C} = {9A989DE7-89B0-4761-83FA-CE8AF8DEEFDB}
 		{CF582C5B-EAA5-4368-BC10-33C1584DB1EA} = {7A93BF5E-D7A6-4D8A-AC22-9D5AF09C4880}
 		{3B2120BB-437C-4485-A039-EF4C0CD5D94D} = {7A93BF5E-D7A6-4D8A-AC22-9D5AF09C4880}
+		{93EFB7C0-9A60-4C33-8415-863E5F4FA9A4} = {D6F2860A-2B15-465E-B1BB-B8B3AE829A9C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CD5C67DB-70F8-4C47-A3F4-7C94BECE1657}
